+++ packages/components-react/src/pages/index.tsx
@@ -165,6 +165,7 @@ import { PopoverBottomPage } from './generated/PopoverBottom';
 import { PopoverCompactPage } from './generated/PopoverCompact';
 import { PopoverCssVariablesPage } from './generated/PopoverCssVariables';
 import { PopoverEdgeCasePage } from './generated/PopoverEdgeCase';
+import { PopoverFlexParentPage } from './generated/PopoverFlexParent';
 import { PopoverLeftPage } from './generated/PopoverLeft';
 import { PopoverRightPage } from './generated/PopoverRight';
 import { PopoverScrolledPage } from './generated/PopoverScrolled';
@@ -997,6 +998,11 @@ export const generatedRoutes: RouteType[] = [
     name: 'Popover Edge Case',
     path: '/popover-edge-case',
     element: <PopoverEdgeCasePage />,
+  },
+  {
+    name: 'Popover Flex Parent',
+    path: '/popover-flex-parent',
+    element: <PopoverFlexParentPage />,
   },
   {
     name: 'Popover Left',